services/graphAssistantApp/requirements.txt
fastapi
uvicorn
python-dotenv
neo4j
pandas
pydantic
google-genai
requests
